AS7-5078 cli if statements parsing: a bit of refactoring, extracted tests to a separate file
